Test invalid switch
